refactor: Separate navbar components
