+++ src/components/utils/data.js
@@ -1,0 +1,16 @@
+export const dropdownData = {
+    applications: [
+        {
+            title: "new application",
+            to: "/application",
+            state: { type: "new" },
+        },
+        { title: "renewal", to: "/application", state: { type: "renewal" } },
+        { title: "re-Issue", to: "/application", state: { type: "re-issue" } },
+    ],
+    contact: [
+        { title: "support center", to: "/support" },
+        { title: "capture centers", to: "/capture-centers" },
+        { title: "driving schools", to: "/driving-schools" },
+    ],
+};
